ndoo/requirements.txt
Django==1.10.3
djangorestframework==3.5.2
psycopg2==2.6.2
selenium==3.0.1
